upload e download imagem

+++ src/main/java/org/woof/woofjoybackend/s3connect/controller/StorageController.java
@@ -1,13 +1,11 @@
 package org.woof.woofjoybackend.s3connect.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.woof.woofjoybackend.configuration.security.jwt.GerenciadorTokenJwt;
-import org.woof.woofjoybackend.domain.entity.Usuario;
+import org.woof.woofjoybackend.domain.entity.DonoImagem;
 import org.woof.woofjoybackend.s3connect.service.StorageService;
 import org.woof.woofjoybackend.service.users.ServiceUser;
 
@@ -20,7 +18,7 @@ public class StorageController {
 
     private final StorageService service;
     private final ServiceUser serviceUser;
-    private final GerenciadorTokenJwt jwtTokenManager;
+
 
     @GetMapping("/connection")
     public ResponseEntity<Boolean> connectionHealthCheck() {
@@ -30,28 +28,33 @@ public class StorageController {
 
     @PostMapping("/upload/{tipo}")
     public ResponseEntity<String> uploadProfileImg(@PathVariable String tipo, @RequestParam(value = "file") MultipartFile file, @RequestHeader("Authorization") String bearerToken) {
-        String emailDono = jwtTokenManager.getUsernameFromToken(bearerToken.substring(7));
-        Usuario dono = serviceUser.getByEmail(emailDono);
-
+        DonoImagem dono = serviceUser.getDonoByToken(bearerToken);
+        if (dono == null) {
+            return ResponseEntity.notFound().build();
+        }
         if (tipo.equalsIgnoreCase("perfil")) {
-            return new ResponseEntity<>(service.uploadProfileImg(file, dono.getDonoImagem().getId()), HttpStatus.OK);
-        } else if (tipo.equalsIgnoreCase("img")){
-            return new ResponseEntity<>(service.uploadImg(file, dono.getDonoImagem().getId()), HttpStatus.OK);
+            return new ResponseEntity<>(service.uploadProfileImg(file, dono.getId()), HttpStatus.OK);
+        } else if (tipo.equalsIgnoreCase("img")) {
+            return new ResponseEntity<>(service.uploadImg(file, dono.getId()), HttpStatus.OK);
         }
         return ResponseEntity.badRequest().build();
     }
 
     @GetMapping("/url/{tipo}")
     public ResponseEntity<List<String>> getImgUrl(@PathVariable String tipo, @RequestHeader("Authorization") String bearerToken) {
-        String emailDono = jwtTokenManager.getUsernameFromToken(bearerToken.substring(7));
-        Usuario dono = serviceUser.getByEmail(emailDono);
-        return ResponseEntity.ok().body(service.getImgUrl(dono.getDonoImagem().getId(), tipo));
+        DonoImagem dono = serviceUser.getDonoByToken(bearerToken);
+        if (dono == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok().body(service.getImgUrl(dono.getId(), tipo));
     }
 
     @DeleteMapping("/delete/{fileName}")
     public ResponseEntity<String> deleteFile(@PathVariable String fileName, @RequestHeader("Authorization") String bearerToken) {
-        String emailDono = jwtTokenManager.getUsernameFromToken(bearerToken.substring(7));
-        Usuario dono = serviceUser.getByEmail(emailDono);
+        DonoImagem dono = serviceUser.getDonoByToken(bearerToken);
+        if (dono == null) {
+            return ResponseEntity.notFound().build();
+        }
         return new ResponseEntity<>(service.deleteFile(fileName, dono.getId()), HttpStatus.OK);
     }
 }
